isolated heartbeat java agent to a separate jar
