--- flask/launcher/src/lockFileTest/java/net/corda/flask/launcher/LockFileTestMain.java
@@ -1,21 +1,0 @@
-package net.corda.flask.launcher;
-
-import lombok.SneakyThrows;
-
-import java.nio.file.Path;
-import java.nio.file.Paths;
-
-public class LockFileTestMain {
-
-    @SneakyThrows
-    public static void main(String[] args) {
-        Path lockfilePath = Paths.get(args[0]);
-        boolean shared = Boolean.parseBoolean(args[1]);
-        boolean keep = Boolean.parseBoolean(args[2]);
-        try (LockFile lockfile = LockFile.acquire(lockfilePath, shared)) {
-            while (keep) {
-                Thread.sleep(1000);
-            }
-        }
-    }
-}
