+++ settings.gradle
@@ -27,8 +27,9 @@ include 'cordapp'
 include 'jar-filter'
 include 'jar-filter:unwanteds'
 include 'flask'
-include 'flask:common'
-include 'flask:launcher'
+include 'flask:flask-common'
+include 'flask:flask-heartbeat-agent'
+include 'flask:flask-launcher'
 
 gradleEnterprise {
     buildScan {
